+++ src/denomination_functions.cpp
@@ -125,7 +125,7 @@ std::vector<CZerocoinMint> getSpends(
 // -------------------------------------------------------------------------------------------------------
 void listSpends(const std::vector<CZerocoinMint>& vSelectedMints)
 {
-    std::map<libzerocoin::CoinDenomination, unsigned int> mapZerocoinSupply;
+    std::map<libzerocoin::CoinDenomination, int64_t> mapZerocoinSupply;
     for (auto& denom : libzerocoin::zerocoinDenomList)
         mapZerocoinSupply.insert(std::make_pair(denom, 0));
